puled out translators
