+++ src/index.ts
@@ -12,5 +12,4 @@ export * from './projects';
 export * from './blocks';
 export * from './drafts';
 export * from './comments';
-export * from './translators';
 export * from './templates';
